refactor: extract the grouped events view
